fixed int32 error in PhoneNo use Get<long> instead of Get

+++ Models/Details.cs
@@ -5,7 +5,7 @@
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
-        public double PhoneNo { get; set; }
+        public long PhoneNo { get; set; }
         public string? ProfilePhoto { get; set; }
         public string? Description { get; set; }
     }
